+++ CHANGELOG.md
@@ -5,6 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 [//]: # (## [Unreleased])
 
+## [2026.05] - 2026-05-02
+
+### Fixed
+
+- Free change PIN dialog on click outside the dialog. This fixes it not
+  opening again after closing in such a way.
+
+### Added
+
+- [#48](https://github.com/oleasteo/koreader-screenlockpin/issues/48)
+  Support the new "Also delete plugin settings" option on plugin disable.
+- Keep menu open while changing the PIN.
+
 ## [2026.02] - 2026-02-08
 
 Minor update with a fix for transparent wallpaper support. Also, I've added an